LayoutTests/plugins/unsupported-plugin-expected.txt

CONSOLE MESSAGE: Tried to use an unsupported plug-in.
A console log message should appear stating the plugin is unsupported.



PASS Testing unsupported plug-in
